Move assistant builder and direct chat data classes into their own files
