+++ app/MindWork AI Studio/Assistants/Builder/AssistantBuilderPluginMetadata.cs
@@ -1,0 +1,8 @@
+namespace AIStudio.Assistants.Builder;
+
+internal sealed class AssistantBuilderPluginMetadata
+{
+    public string Name { get; init; } = string.Empty;
+    public string Description { get; init; } = string.Empty;
+    public string[] Categories { get; init; } = [];
+}
