+++ app/MindWork AI Studio/Tools/Services/AssistantBuilderChatLaunchRequest.cs
@@ -1,0 +1,3 @@
+namespace AIStudio.Tools.Services;
+
+public sealed record AssistantBuilderChatLaunchRequest(string WorkspaceName, string? ProviderId, string? ProfileId, string? ChatTemplateId, IReadOnlyList<string>? DataSourceIds);
