+++ app/MindWork AI Studio/Tools/Services/AssistantPluginDraftGenerationResult.cs
@@ -1,0 +1,3 @@
+namespace AIStudio.Tools.Services;
+
+public sealed record AssistantPluginDraftGenerationResult(bool Success, string Markdown, string Issue);
