+++ app/MindWork AI Studio/Tools/Services/AssistantPluginGenerationDraft.cs
@@ -1,0 +1,3 @@
+namespace AIStudio.Tools.Services;
+
+public sealed record AssistantPluginGenerationDraft(bool Success, string Lua, string PluginName, string Issue);
