+++ app/MindWork AI Studio/Tools/Services/AssistantPluginLuaGenerationRequest.cs
@@ -1,0 +1,7 @@
+namespace AIStudio.Tools.Services;
+
+public sealed record AssistantPluginLuaGenerationRequest(
+    Guid PluginId,
+    string ApprovedAssistantDraft,
+    string ReviewNotes,
+    AssistantBuilderChatLaunchRequest? ChatLaunch);
