+++ app/MindWork AI Studio/Tools/Services/AssistantPluginRevisionDraft.cs
@@ -1,0 +1,3 @@
+namespace AIStudio.Tools.Services;
+
+public sealed record AssistantPluginRevisionDraft(bool Success, string Lua, string PluginName, string Issue);
