+++ app/MindWork AI Studio/Tools/Services/DirectChatStartResult.cs
@@ -1,0 +1,5 @@
+using AIStudio.Chat;
+
+namespace AIStudio.Tools.Services;
+
+public sealed record DirectChatStartResult(ChatStartRequest? Request, string ErrorMessage);
